Prepare repo to host more than one Array type
The README is restructured to describe all of the different data
containers types.
The Makefile is modified to given other names to output files:
th final library is renamed to `libarrays.a`, the test executable is
renamed to `test_arrays`.
The test suite for the *Array* structure is renamed to `test_array()`
and returns `void`, it is called from the global test file `test.c`
which holds the definition of *main()*.