Renamed all functions to be clearer.
Though it loses in elegance, it is more understandable. Each function
has the `a` prefix (indicating that is concerns the *Array* type)
separated from the base name of the function (that correspond to the
action performed by the function) with a single underscore:

`afree` becomes `a_free`, `aprintf` becomes `a_printf`, etc.

Moreover this will prevent some potential name clashes with standard
functions for future types.