Use flexible array member
The double pointer to store the elements is replaced with a flexible
array member, that is an array defined without a size that is sized
when the containing structure is allocated.
There are limitations though: the structure has to be allocated
dynamically (*malloc()* n' stuff) and the array must be the last member
in the definition of the structure.