Add bounds crossing guards
The functions fa_get() and fa_set() were not protected against
illegal access (index > size).

Now fa_get() will return `NULL` on invalid index (beware that it is
also a valid value for an actual element) and fa_set() will not act
when it is called with an invalid index.