Change implementation of Array to internally use a FixedArray
The FixedArray member contains the data for the array's capacity and
elements, the array itself contains only the information about its own
size (the number of elements that are actually set)

When the array overflows, the fixed member si reallocated to a larger
copy of itself (after being freed).

Note that this implementation loses the size of a `long` in memory
(eight bytes) due to struct padding.