Fix behavouir of a_add()
The function would not try to validate an index strictly greater than
the size of the array, although those values should be (in)validated;
this is fixed by changing the comparison operation to a *not equal to*
operator (because, as said in explanatory comment, an index of value
the size of the array is a valid value).

Also added (commented out) debug messages in valid() to indicate in
which branch the control flow passes.