Fix ladd
The function was horribly flawed.
The validsation of the index was not made at first, but after the
instantiation of a new node. If the item is added at the head of the
list (ie. index 0) the member *next* of the added node was left set to
`NULL` leading to systematic segmentation faults. At the end of the
function, the *len* was post-incremented instead of the equivalent but
more efficient pre-increment.