Remove support for negative indices
The negative indices allowed to access elements from an *Array* or a
*LinkedList* counting from the end, instead of from the beginning of
the structure, `-1` being the last element and `-SIZE` the first
(`SIZE` being the actual number of elements in the structure, as
returned by `a_size` or `ll_len`) -- exactly as Python's negative
indices.

This functionality was never tested, though it did not add much
overhead to a function call and did not represent many lines of code,
it is rarely a useful case, so its support is completely removed.

Remember:

> Perfection is achieved, not when there is nothing more to add, but
> when there is nothing left to take away
!