Revise functions ll_add and ll_drop
The implementations are now shorter (in terms of lines of code).
Some `if` statements are merged together, all variables declarations
are made at the top of each function, made use of the *plug* pattern
to hold in a pointer a value depending on some conditions and act on
the value by dereferencing the pointer.