Use errno
Deinitions of functions that might be in an error state now use the
*errno* variable to specify their status.
Concerned functions are constructor-like functions (the error being a
memory allocation failure, or invalid argument if specified size for
the structure is `0`), getters and setters (the error status being an
invalid index given), etc.
For these functions, the variable is assured to be set to `0` if the
function returned correctly, and to a non-zero specific value otherwise.