Add doc for errno
Every function of the types FixedArray, Array and LinkedList that uses
the error variable *errno* is explicitly documented what possible
values are given to the variable, and the circumstances where the
variable would be set to these values.

Also adds an inline comment in a source file to explicit the fact that
setting errno here is not necessary.