Remove dependency of install make rule
The rule built the project if it was not, but with root permissions --
then the archive and object files belonged to root and had to be
cleaned with sudo. The dependency is removed so that the rule won't try
to build the lib anymore; it is expected to be explicitly built
beforeheand.