Rewrite ll_cond and test
The function now takes an extra argument that correspond to a value to
test each element of the list against. The value is tested through the
provided function (as argument) that takes now two arguments: the
element of the list, and the test value.

Two tests are added for the function: one with a valid value and one
without. The tests check the return value and errno.