Update project architecture & Makefile
The test sources are moved to a new directory, `test` and the prefix
`test_` is removed (though it is kept for object files for
disambiguation).

The Makefile is improved -- variables `CPPFLAGS` (options to the
preprocessor), `TEST_DIR` (directory with test sources) are added, link
to the Clog library is fixed (`-lclog`).