Use CUTE facility
The tests are now implemented using the library *CUTE*.
Unit tests are now truly independent and can be run non-sequentially.
Provided assertions are used too.

The way a test case could be given its tests turned out to be not so
obvious: the options were either to run the cases independently -- and
losing the point of the test suite -- or to include the test sources,
and in doing so I would have to rename all the internal variables and
functions.
The actual fashion I opted for was to implemented a function for each
structure, called `build_case_*`, that would internally instantiate its
associate case and populate it with its tests. These builder functions
are then called from `main()`, and are the only functions declared
externally in the main test source, and no conflict or name clash
arose.