Use standard C99 type size_t and POSIX ssize_t
All occurences of `unsigned int` for array index or length are replaced
with `size_t`, as this type is more suited for these uses. (moreover it
is a wider type, so the arrays can contain even *more* elements!)
All occurences of `int` as an index or error value (usually `-1`) are
replaced with `ssize_t`, the POSIX-defined equivalent signed type --
actually, it is only ensured to contain *at least* integers from the
range `[-1, 2^15 - 1]`, but it is usually the signed counterpart of
`size_t`.