Fix tests
Replace accesses to array index with `+` for address with `&[]`, remove
variables `param` of type `data_t*` with direct addresses of the
variable it takes the address from, replace addresses in log messages
with the actual integer value, preceded with `&` to indicate address
(e.g. `&(42)` for the address of a variable whose integer value is
`42`).