Declare a new structure: ArrayMap
The structure binds two kinds of elements, keys and values, in a
one-to-one relationship (each value has its distinct key).
The elements are stored in two parallel arrays, where the index of a
value matches that of its key. The keys are stored in a `SortedArray`
for a faster retrieval algorithm (binary search => `O(log n)`).